+++ factual/query/__init__.py
@@ -5,3 +5,4 @@ from .submit import Submit
 from .facets import Facets
 from .flag import Flag
 from .geopulse import Geopulse
+from .geocode import Geocode