+++ CHANGELOG.md
@@ -3,3 +3,4 @@
  * added changelog
  * added parameter dict support to raw_read
  * added geopulse feature
+ * added reverse geocode feature